projeto finalizado

+++ src/hook_move/controls/rotate_left.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   rotate_left.c                                      :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: rbutzke <rbutzke@student.42sp.org.br>      +#+  +:+       +#+        */
+/*   By: rbutzke <rbutzke@student.42so.org.br>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/08/20 08:10:30 by rbutzke           #+#    #+#             */
-/*   Updated: 2024/08/23 07:47:07 by rbutzke          ###   ########.fr       */
+/*   Updated: 2024/09/05 16:11:35 by rbutzke          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,36 +20,6 @@
 static void	rotate_dir(t_data *data);
 static void	rotate_plane(t_data *data);
 
-/**
- * @brief Updates the player's direction and camera plane when rotating to the left.
- * 
- * This function handles the logic needed to rotate the player's view to the left.
- * When called, it updates both the direction vector and the camera plane,
- * ensuring the player's perspective in the game world is correctly adjusted.
- * 
- * @param data A pointer to the main game data structure (t_data).
- *  
- * The following attributes are updated within this function:
- *  - data->coord->dir[X]:
- * 		Controls the horizontal component of the player's direction vector,
- *		which affects the player's viewing angle along the east-west axis.
- *  - data->coord->dir[Y]:
- *      Controls the vertical component of the player's direction vector,
- * 		which affects the player's viewing angle along the north-south axis.
- *  - data->coord->plane[](Y or X):
- *      Define the camera plane, which is perpendicular to the direction vector.
- * 		This plane is crucial for determining the field of view (FOV)
- * 		and correctly rendering the 3D perspective on the 2D screen.
- * 
- * The auxiliary functions:
- * 	- rotate_dir(data):
- * 		Calculates and updates the direction vector.
- *	- rotate_plane(data);
- * 		Calculates and updates the plane vector.
- * 
- * @return int The return is to signal an update in the
- * 			   coordinates and execute the updated rendering.
- */
 int	rotate_cam_left(t_data *data)
 {
 	rotate_dir(data);
@@ -64,29 +34,30 @@ int	rotate_cam_left(t_data *data)
  * changing the angle at which the player is facing. It effectively rotates
  * the player's view within the game world, adjusting the direction in which
  * the player is looking.
- * 
- * 
+ *
  * @param data A pointer to the main game data structure (t_data).
- * 
+ *
  * To update, we first create some temporary variables:
- * 
  * 	- old_dir_y:
- * 		Is the value of Y before the update, (it will be used to calculate the X axis).
- * 	- cos_dir[] and sin_dir[]:
- * 		Is the multiplication of the corresponding axis (Y or X) by the cosine and sine of ROT_SPEED.
- * 
+ * 	Is the value of Y before the update,
+ *  (it will be used to calculate the X axis).
+ *
+ * - cos_dir[] and sin_dir[]:
+ * 	Is the multiplication of the corresponding axis (Y or X)
+ *  by the cosine and sine of ROT_SPEED.
+ *
  * Then the Y and X axis of the direction are updated: 
  * 	- data->coord->dir[Y]:
- * 		With the subtraction between cos_dir[Y] - sin_dir[X].
+ * 	With the subtraction between cos_dir[Y] - sin_dir[X].
+ * 
  * 	- data->coord->dir[X]:
- * 		With the sum between sin_dir[Y] + cos_dir[X].
+ * 	With the sum between sin_dir[Y] + cos_dir[X].
  **/
-
 static void	rotate_dir(t_data *data)
 {
-	double	old_dir_y;
-	double	cos_dir[2];
-	double	sin_dir[2];
+	long double	old_dir_y;
+	long double	cos_dir[2];
+	long double	sin_dir[2];
 
 	old_dir_y = data->coord->dir[Y];
 	sin_dir[Y] = old_dir_y * sin(ROT_SPEED);
@@ -110,9 +81,11 @@ static void	rotate_dir(t_data *data)
  * To update, we first create some temporary variables:
  * 
  * 	- old_plane_y:
- * 		Is the value of Y before the update, (it will be used to calculate the X axis).
+ *  Is the value of Y before the update,
+ *  (it will be used to calculate the X axis).
  * 	- cos_plane[] and sin_plane[2]:
- * 		Is the multiplication of the corresponding axis (Y or X) by the cosine and sine of ROT_SPEED.
+ * 	Is the multiplication of the corresponding axis (Y or X)
+ *  by the cosine and sine of ROT_SPEED.
  * 
  * Then the Y and X axis of the direction are updated: 
  * 	- data->coord->plane[Y]:
@@ -122,9 +95,9 @@ static void	rotate_dir(t_data *data)
  * */
 static void	rotate_plane(t_data *data)
 {
-	double	cos_plane[2];
-	double	sin_plane[2];
-	double	old_plane_y;
+	long double	cos_plane[2];
+	long double	sin_plane[2];
+	long double	old_plane_y;
 
 	old_plane_y = data->coord->plane[Y];
 	sin_plane[Y] = old_plane_y * sin(ROT_SPEED);
