+++ src/hook_move/controls/rotate_right.c
@@ -3,47 +3,23 @@
 /*                                                        :::      ::::::::   */
 /*   rotate_right.c                                     :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: rbutzke <rbutzke@student.42sp.org.br>      +#+  +:+       +#+        */
+/*   By: rbutzke <rbutzke@student.42so.org.br>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/08/20 08:10:30 by rbutzke           #+#    #+#             */
-/*   Updated: 2024/08/23 07:48:52 by rbutzke          ###   ########.fr       */
+/*   Updated: 2024/09/05 16:11:35 by rbutzke          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include <MLX42.h>
 #include "data.h"
 #include "defines.h"
-#include <math.h>
-#include "render_ceiling_floor.h"
 #include "ray_casting.h"
+#include "render_ceiling_floor.h"
+#include <MLX42.h>
+#include <math.h>
 
 static void	rotate_dir(t_data *data);
 static void	rotate_plane(t_data *data);
 
-/**
- * @brief Updates the player's direction and camera plane when rotating to the right.
- * 
- * This function handles the logic needed to rotate the player's view to the right.
- * When called, it updates both the direction vector and the camera plane,
- * ensuring the player's perspective in the game world is correctly adjusted.
- * 
- * @param data A pointer to the main game data structure (t_data).
- *  
- * The following attributes are updated within this function:
- *  - data->coord->dir[X]:
- * 		Controls the horizontal component of the player's direction vector,
- *		which affects the player's viewing angle along the east-west axis.
- *  - data->coord->dir[Y]:
- *  	Controls the vertical component of the player's direction vector,
- * 		which affects the player's viewing angle along the north-south axis.
- *  - data->coord->plane[](Y or X):
- *   	Define the camera plane, which is perpendicular to the direction vector.
- * 	 	This plane is crucial for determining the field of view (FOV)
- * 	 	and correctly rendering the 3D perspective on the 2D screen.
- * 
- * @return int The return is to signal an update in the
- * 			   coordinates and execute the updated rendering.
- */
 int	rotate_cam_right(t_data *data)
 {
 	rotate_dir(data);
@@ -53,22 +29,25 @@ int	rotate_cam_right(t_data *data)
 
 /**
  * @brief Rotates the player's direction vector to the right.
- * 
+ *
  * This function applies a rotation matrix to the player's direction vector,
  * changing the angle at which the player is facing. It effectively rotates
  * the player's view within the game world, adjusting the direction in which
  * the player is looking.
- * 
+ *
  * @param data A pointer to the main game data structure (t_data).
- * 
+ *
  * To update, we first create some temporary variables:
- * 
+ *
  * 	- old_dir_y:
- * 		Is the value of Y before the update, (it will be used to calculate the X axis).
+ *   Is the value of Y before the update,
+ *   (it will be used to calculate the X axis).
+ *
  * 	- cos_dir[] and sin_dir[]:
- * 		Is the multiplication of the corresponding axis (Y or X) by the cosine and sine of -ROT_SPEED.
- * 
- * Then the Y and X axis of the direction are updated: 
+ * 	Is the multiplication of the corresponding axis (Y or X) 
+ *  by the cosine and sine of -ROT_SPEED.
+ *
+ * Then the Y and X axis of the direction are updated:
  * 	- data->coord->dir[Y]:
  * 		With the subtraction between cos_dir[Y] - sin_dir[X].
  * 	- data->coord->dir[X]:
@@ -76,9 +55,9 @@ int	rotate_cam_right(t_data *data)
  **/
 static void	rotate_dir(t_data *data)
 {
-	double	old_dir_y;
-	double	cos_dir[2];
-	double	sin_dir[2];
+	long double	old_dir_y;
+	long double	cos_dir[2];
+	long double	sin_dir[2];
 
 	old_dir_y = data->coord->dir[Y];
 	sin_dir[Y] = old_dir_y * sin(-ROT_SPEED);
@@ -91,32 +70,36 @@ static void	rotate_dir(t_data *data)
 
 /**
  * @brief Rotates the player's camera plane to the left.
- * 
+ *
  * This function adjusts the player's camera plane, which is necessary
  * for maintaining the correct field of view after the player's direction
  * has been rotated. The camera plane defines the part of the world that
  * is visible on the screen and is always perpendicular to the direction vector.
- * 
+ *
  * @param data A pointer to the main game data structure (t_data).
- * 
+ *
  * To update, we first create some temporary variables:
+ *	- old_plane_y:
+ * 	Is the value of Y before the update,
+ * 	(it will be used to calculate the X axis).
  * 
- * 	- old_plane_y:
- * 		Is the value of Y before the update, (it will be used to calculate the X axis).
  * 	- cos_plane[] and sin_plane[2]:
- * 		Is the multiplication of the corresponding axis (Y or X) by the cosine and sine of -ROT_SPEED.
+ *  Is the multiplication of the corresponding axis (Y or X)
+ *  by the cosine and sine of -ROT_SPEED.
+ *
+ * Then the Y and X axis of the direction are updated:
  * 
- * Then the Y and X axis of the direction are updated: 
  * 	- data->coord->plane[Y]:
  * 		With the subtraction between cos_dir[Y] - sin_dir[X].
+ * 
  * 	- data->coord->plane[X]:
  * 		With the sum between sin_dir[Y] + cos_dir[X].
  * */
 static void	rotate_plane(t_data *data)
 {
-	double	cos_plane[2];
-	double	sin_plane[2];
-	double	old_plane_y;
+	long double	cos_plane[2];
+	long double	sin_plane[2];
+	long double	old_plane_y;
 
 	old_plane_y = data->coord->plane[Y];
 	sin_plane[Y] = old_plane_y * sin(-ROT_SPEED);
